==> notebooks/01_data_collection.ipynb <==
import pandas as pd
import os
import glob

dosya_deseni = 'C:/Users/yunus/Desktop/Bahar Dersleri/Yapay Zeka Proje/Araç_Arıza_Eşleştirme/data/raw/carcomplaints*.csv'
dosya_listesi = glob.glob(dosya_deseni)

veri_cerceveleri = []
for dosya_adi in dosya_listesi:
    try:
        df = pd.read_csv(dosya_adi)
        veri_cerceveleri.append(df)
        print(f"{dosya_adi} başarıyla okundu")
    except Exception as e:
        print(f"{dosya_adi} okunurken hata oluştu: {e}")
if veri_cerceveleri:
    birlesik_veri = pd.concat(veri_cerceveleri, ignore_index=True)

    # Birleştirilmiş veriyi görüntüleyin (isteğe bağlı)
    print("\nBirleştirilmiş Veri:")
    print(birlesik_veri.head()) # İlk birkaç satırı gösterir
    print(f"\nToplam {len(birlesik_veri)} satır birleştirildi.")

    # Birleştirilmiş veriyi yeni bir CSV dosyasına kaydedin (isteğe bağlı)
    birlesik_dosya_adi = 'tum_carcomplaints.csv'
    birlesik_veri.to_csv(birlesik_dosya_adi, index=False, encoding='utf-8')
    print(f"\nBirleştirilmiş veri '{birlesik_dosya_adi}' olarak kaydedildi.")
else:
    print("\nBelirtilen desene uygun hiçbir dosya bulunamadı.")
