==> notebooks/02_data_preprocess.ipynb <==
import pandas as pd
import os
import numpy as np
import nltk
from nltk.tokenize import word_tokenize, sent_tokenize
from nltk.corpus import stopwords
from nltk.stem import WordNetLemmatizer, PorterStemmer
import re

df = pd.read_csv("C:/Users/yunus/Desktop/Bahar Dersleri/Yapay Zeka Proje/arac-arizadeseni-eslestirme/data/raw/all_data.csv")

df.head()

def preprocess_text(text):
    sentences = nltk.sent_tokenize(text)
    processed_sentences = []
    for sentence in sentences:
        # 2. Küçük harfe çevirme
        sentence = sentence.lower()
        
        # 3. Noktalama işaretlerini kaldırma
        sentence = re.sub(r'[^\w\s]', '', sentence)
        
        # 4. Tokenizasyon
        tokens = nltk.word_tokenize(sentence)
        
        # 5. Stop word temizliği
        stop_words = set(stopwords.words('turkish'))
        tokens = [token for token in tokens if token not in stop_words]
        
        # 6. Lemmatizasyon
        lemmatizer = WordNetLemmatizer()
        tokens = [lemmatizer.lemmatize(token) for token in tokens]
        
        processed_sentences.append(tokens)
    
    return processed_sentences


# Fonksiyonu kullanarak işlemleri uygulama
df['comments_processed'] = df['comments'].apply(preprocess_text)

print(df['comments_processed'].head())

df.to_csv("C:/Users/yunus/Desktop/Bahar Dersleri/Yapay Zeka Proje/arac-arizadeseni-eslestirme/data/raw/preprocessed_data_lemmatized_only.csv", index=False)

veri = pd.read_csv("C:/Users/yunus/Desktop/Bahar Dersleri/Yapay Zeka Proje/arac-arizadeseni-eslestirme/data/raw/preprocessed_data_lemmatized_only.csv")
veri.head(10)

df2 = pd.read_csv("C:/Users/yunus/Desktop/Bahar Dersleri/Yapay Zeka Proje/arac-arizadeseni-eslestirme/data/raw/all_data.csv")

def stem_text(text):
  #Cümlelere ayırma
    sentences = nltk.sent_tokenize(text)
    
    processed_sentences = []
    for sentence in sentences:
        #Küçük harfe çevirme
        sentence = sentence.lower()
        #Noktalama işaretlerini kaldırma
        sentence = re.sub(r'[^\w\s]', '', sentence)
        #Tokenizasyon
        tokens = nltk.word_tokenize(sentence)
        #Stop word temizleme
        stop_words = set(stopwords.words('english'))
        tokens = [token for token in tokens if token not in stop_words]
        #Stemming
        stemmer = PorterStemmer()
        tokens = [stemmer.stem(token) for token in tokens]
        processed_sentences.append(tokens)
    return processed_sentences

df2['comments_stemmed'] = df2['comments'].apply(stem_text)

df2.to_csv("C:/Users/yunus/Desktop/Bahar Dersleri/Yapay Zeka Proje/arac-arizadeseni-eslestirme/data/raw/preprocessed_data_stemmed_only.csv",index=False)

veri2 = pd.read_csv("C:/Users/yunus/Desktop/Bahar Dersleri/Yapay Zeka Proje/arac-arizadeseni-eslestirme/data/processed/preprocessed_data_stemmed_only.csv")

veri2.head()
