==> notebooks/03_vectorization_tf-idf.ipynb <==
import pandas as pd
import numpy as np
from sklearn.feature_extraction.text import TfidfVectorizer
from sklearn.metrics.pairwise import cosine_similarity

df_lemmatized = pd.read_csv("C:/Users/yunus/Desktop/Bahar Dersleri/Yapay Zeka Proje/arac-arizadeseni-eslestirme/data/processed/preprocessed_data_lemmatized_only.csv")
df_stemmed = pd.read_csv("C:/Users/yunus/Desktop/Bahar Dersleri/Yapay Zeka Proje/arac-arizadeseni-eslestirme/data/processed/preprocessed_data_stemmed_only.csv")

# TF İDF Vektörizasyon işlemi
tfidf_vectorizer_lemma = TfidfVectorizer()
tfidf_matrix_lemma = tfidf_vectorizer_lemma.fit_transform(df_lemmatized['comments_processed'])
tfidf_df_lemma = pd.DataFrame(tfidf_matrix_lemma.toarray(), columns=tfidf_vectorizer_lemma.get_feature_names_out())

tfidf_vectorizer_stem = TfidfVectorizer()
tfidf_matrix_stem = tfidf_vectorizer_stem.fit_transform(df_stemmed['comments_stemmed'])
tfidf_df_stem = pd.DataFrame(tfidf_matrix_stem.toarray(), columns=tfidf_vectorizer_stem.get_feature_names_out())

#Dosyaların kaydedilmesi
tfidf_df_lemma.to_csv("C:/Users/yunus/Desktop/Bahar Dersleri/Yapay Zeka Proje/arac-arizadeseni-eslestirme/data/processed/tfidf_lemmatized.csv", index=False)
tfidf_df_stem.to_csv("C:/Users/yunus/Desktop/Bahar Dersleri/Yapay Zeka Proje/arac-arizadeseni-eslestirme/data/processed/tfidf_stemmed.csv", index=False)

# Lemmatize edilmiş veri için
first_sentence_vector_lemma = tfidf_df_lemma.iloc[0]
top_5_words_lemma = first_sentence_vector_lemma.sort_values(ascending=False).head(5)

print("\nLemmatize edilmiş veri - İlk cümlenin en yüksek TF-IDF skorlu 5 kelimesi:\n")
print(top_5_words_lemma)

# Stemmed veri için
first_sentence_vector_stem = tfidf_df_stem.iloc[0]
top_5_words_stem = first_sentence_vector_stem.sort_values(ascending=False).head(5)

print("\nStemmed veri - İlk cümlenin en yüksek TF-IDF skorlu 5 kelimesi:\n")
print(top_5_words_stem)

# Lemmatize edilmiş veri için
cosine_sim_matrix_lemma = cosine_similarity(tfidf_matrix_lemma)
cosine_sim_df_lemma = pd.DataFrame(cosine_sim_matrix_lemma)

print("\nLemmatize edilmiş veri - Kosinüs Benzerliği Matrisi:\n")
print(cosine_sim_df_lemma.head())

# Stemmed veri için
cosine_sim_matrix_stem = cosine_similarity(tfidf_matrix_stem)
cosine_sim_df_stem = pd.DataFrame(cosine_sim_matrix_stem)

print("\nStemmed veri - Kosinüs Benzerliği Matrisi:\n")
print(cosine_sim_df_stem.head())
